+++ packages/solid/src/components/toast/toaster.tsx
@@ -23,13 +23,13 @@ export const Toaster = (props: ToasterProps) => {
     store: toasterProps.toaster,
     id: createUniqueId(),
     dir: locale()?.dir,
-    getRootNode: () => env()?.getDocument(),
+    getRootNode: env()?.getRootNode,
   }))
 
   const api = createMemo(() => toast.group.connect(service, normalizeProps))
   const toasts = createMemo(() => api().getToasts())
 
-  const mergedProps = mergeProps(api().getGroupProps(), localProps)
+  const mergedProps = mergeProps(() => api().getGroupProps(), localProps)
 
   return (
     <ark.div {...mergedProps}>
@@ -55,7 +55,7 @@ const ToastActor = (props: ToastActorProps) => {
   const localProps = createMemo(() => ({
     ...props.value(),
     parent: props.parent,
-    index: props.index,
+    index: props.index(),
   }))
   const service = useMachine(toast.machine, localProps)
   const api = createMemo(() => toast.connect(service, normalizeProps))