+++ packages/solid/CHANGELOG.md
@@ -13,6 +13,7 @@ description: All notable changes will be documented in this file.
 ### Fixed
 
 - **Presence**: Fixed issue where `onExitComplete` was not being called.
+- **Toaster**: Fixed issue where `--index` was not applied correctly.
 
 ## [5.1.1] - 2025-03-17